requirements.txt
sseclient-py
urllib3
pytest